window: add client side decorations interface
While the new decorations look the same as the ones before, the
approach in drawing them was changed, and now decorations are more
instant mode, meaning that you create and drop them when needed.

Fixes #273.

+++ src/shell/xdg/window/mod.rs
@@ -258,8 +258,8 @@ impl Window {
         decoration.data::<WindowData>().and_then(|data| data.0.upgrade()).map(Window)
     }
 
-    pub fn show_window_menu(&self, seat: &wl_seat::WlSeat, serial: u32, position: (u32, u32)) {
-        self.xdg_toplevel().show_window_menu(seat, serial, position.0 as i32, position.1 as i32);
+    pub fn show_window_menu(&self, seat: &wl_seat::WlSeat, serial: u32, position: (i32, i32)) {
+        self.xdg_toplevel().show_window_menu(seat, serial, position.0, position.1);
     }
 
     pub fn set_title(&self, title: impl Into<String>) {
